Replace CommentController with CommentsApiController; implement API for listing and deleting comments

+++ src/Controller/Frontend/CommentsController.php
@@ -10,21 +10,24 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 
 #[Route('/comments')]
 final class CommentsController extends AbstractController{
     #[Route(name: 'app_comments_index', methods: ['GET'])]
     public function index(CommentsRepository $commentsRepository): Response
     {
-        return $this->render('frontend/comments/index.html.twig', [
+        return $this->render('comments/index.html.twig', [
             'comments' => $commentsRepository->findAll(),
         ]);
     }
 
     #[Route('/new', name: 'app_comments_new', methods: ['GET', 'POST'])]
+    #[IsGranted('ROLE_USER')]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
         $comment = new Comments();
+        $comment->setUser($this->getUser());
         $form = $this->createForm(CommentsType::class, $comment);
         $form->handleRequest($request);
 
@@ -35,23 +38,29 @@ final class CommentsController extends AbstractController{
             return $this->redirectToRoute('app_comments_index', [], Response::HTTP_SEE_OTHER);
         }
 
-        return $this->render('frontend/comments/new.html.twig', [
+        return $this->render('comments/new.html.twig', [
             'comment' => $comment,
             'form' => $form,
         ]);
     }
 
     #[Route('/{id}', name: 'app_comments_show', methods: ['GET'])]
+    #[IsGranted('ROLE_USER')]
     public function show(Comments $comment): Response
     {
-        return $this->render('frontend/comments/show.html.twig', [
+        return $this->render('comments/show.html.twig', [
             'comment' => $comment,
         ]);
     }
 
     #[Route('/{id}/edit', name: 'app_comments_edit', methods: ['GET', 'POST'])]
+    #[IsGranted('ROLE_USER')]
     public function edit(Request $request, Comments $comment, EntityManagerInterface $entityManager): Response
     {
+        if ($comment->getUser() !== $this->getUser()) {
+            throw $this->createAccessDeniedException('You can only edit your own comments.');
+        }
+
         $form = $this->createForm(CommentsType::class, $comment);
         $form->handleRequest($request);
 
@@ -61,15 +70,20 @@ final class CommentsController extends AbstractController{
             return $this->redirectToRoute('app_comments_index', [], Response::HTTP_SEE_OTHER);
         }
 
-        return $this->render('frontend/comments/edit.html.twig', [
+        return $this->render('comments/edit.html.twig', [
             'comment' => $comment,
             'form' => $form,
         ]);
     }
 
     #[Route('/{id}', name: 'app_comments_delete', methods: ['POST'])]
+    #[IsGranted('ROLE_USER')]
     public function delete(Request $request, Comments $comment, EntityManagerInterface $entityManager): Response
     {
+        if ($comment->getUser() !== $this->getUser()) {
+            throw $this->createAccessDeniedException('You can only delete your own comments.');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$comment->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($comment);
             $entityManager->flush();
